% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{cal_pattern}
\alias{cal_pattern}
\title{Calibrate the relative bank strength to observed conditions}
\usage{
cal_pattern(Q, S, D50, D84, D90, Plot = T, fs = 0)
}
\arguments{
\item{Q}{formative discharge (m3/s)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface (mmm) influencing resistance}

\item{D90}{90th percentile of the bed surface (mm) influencing stability}

\item{Plot}{a logical value that controls the output of summary plots}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{cal_pattern} is a function that calculates the channel dimensions
for a wide range of relative bank strength values, given a set of input
parameters (Q, S, D50, D84, D90, fs). The output data can be used to 
help calibrate the bank strength parameter used to match the model to an
existing river reach
}
