% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{channel_state}
\alias{channel_state}
\title{Calculate the channel state}
\usage{
channel_state(p, y, b, S, D50, D84, fs)
}
\arguments{
\item{p}{bed width (m)}

\item{y}{maximum flow depth for trapezoid (m)}

\item{b}{sideslope angle of the banks (degrees)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface (mmm) influencing resistance}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{channel_state} calculates the characteristics of a trapezoidal channel
with specified dimensions and bed material. The calculated characteristics 
include wetted width (W), mean depth (d), mean velocity (U), total discharge
(Q), shear stress acting on the bed (Tbed), shear stress acting on the bank
(Tbank), volumetric transport capacity (Qb)
}
\details{
#' @param Q formative discharge (m3/s)
}
