% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{est_pattern}
\alias{est_pattern}
\title{Run the regime model for stable channel patterns}
\usage{
est_pattern(Q, mu, S, D50, D84, D90, fs = 0)
}
\arguments{
\item{Q}{formative discharge (m3/s)}

\item{mu}{bank strength, relative to bed (afer Millar, 2005)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface (mmm) influencing resistance}

\item{D90}{90th percentile of the bed surface (mm) influencing stability}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{est_pattern} is a function that runs the UBC Regime Model using
the reach average estimates of surface texture, slope, formative discharge
and relative bank strength. the function calculates the number of anabranches
required for a stable channel pattern, as well as the total channel width (Wt),
mean hydraulic depth (d), mean flow velocity (U), and total volumetric transport
capacity (Qbt)
}
