% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{est_regime}
\alias{est_regime}
\title{Run the regime model in basic mode}
\usage{
est_regime(Q, mu, S, D50, D84, D90, fs = 0)
}
\arguments{
\item{Q}{formative discharge (m3/s)}

\item{mu}{bank strength, relative to bed (afer Millar, 2005)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface (mmm) influencing resistance}

\item{D90}{90th percentile of the bed surface (mm) influencing stability}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{est_regime} is a function that runs the UBC Regime Model using the
reach average estimates of surface texture, slope, formative discharge and
relative bank strength. The function calculates the wetted width (W, in m), 
mean hydraulic depth (d, in m), mean flow velocity (U, in m/s), and the 
sediment transport capacity (Qb, in m3/s)
}
