% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{find_Q}
\alias{find_Q}
\title{Solve for water depth}
\usage{
find_Q(Q, p, b, S, D50, D84, fs)
}
\arguments{
\item{Q}{discharge contained by the channel (m3/s)}

\item{p}{bed width (m)}

\item{b}{sideslope angle of the banks (degrees)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface grain size distribution (mm)}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{find_Q} uses a midpoint solution approach to find water depth that
produces the user-specified stream discharge in a trapezoidal channel with
a known geometry.
}
