% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{find_stable}
\alias{find_stable}
\title{Solve for the stable channel geometry}
\usage{
find_stable(Q, mu, p, S, D50, D84, D90, fs)
}
\arguments{
\item{Q}{formative discharge (m3/s)}

\item{mu}{bank strength, relative to bed (afer Millar, 2005)}

\item{p}{bed width (m)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface (mmm) influencing resistance}

\item{D90}{90th percentile of the bed surface (mm) influencing stability}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{find_stable} solves for the stable bank angle for a given channel 
width, slope, sediment texture, relative bank strength and discharge. The
function returns all of the descriptor variables in \code{channel_state}
}
