% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimeFunctions.R
\name{sim_pattern}
\alias{sim_pattern}
\title{Run a Monte Carlo simulation predicting stable channel characteristics}
\usage{
sim_pattern(Q, mu, S, D50, D84, D90, Uncert = 0.1, Nmc = 100, Plot = T,
  fs = 0)
}
\arguments{
\item{Q}{formative discharge (m3/s)}

\item{mu}{bank strength, relative to bed (afer Millar, 2005)}

\item{S}{energy gradient for the stream reach (m/m)}

\item{D50}{median size of the bed surface (mm) influencing transport rate}

\item{D84}{84th percentile of the bed surface (mmm) influencing resistance}

\item{D90}{90th percentile of the bed surface (mm) influencing stability}

\item{Uncert}{uncertainty for the inputs, expressed as a proportion between 0 and 1.0}

\item{Nmc}{number of Monte Carlo simulations to run (default is 100)}

\item{Plot}{a logical value that controls the output of summary statistics and plots}

\item{fs}{percent of the bed surface that is covered by sand (percent)}
}
\description{
\code{sim_pattern} is a function that launches a Monte Carlo simulation to
estimate the range of channel conditions that could be expected for a given
set of input parameters (Q, mu, S, D50, D84, D90, fs) with a specified level
of uncertainty for the input parameters (expressed as a proportion varying 
from 0 to 1.0)
}
